==> criteo-nips-2017/01-eda.ipynb <==
!ls data/

%load_ext autoreload
%autoreload 2

!zcat data/criteo_train.txt.gz | head

import gzip

import pandas as pd

import numpy as np
import scipy.sparse as sp

from tqdm import tqdm_notebook as tqdm

def parse_features(s):
    split = s.split(' ')
    f0 = split[0]
    assert f0.startswith('0:')
    f0 = int(f0[2:])

    f1 = split[1]
    assert f1.startswith('1:')
    f1 = int(f1[2:])

    idx = []
    values = []
    
    for fv in split[2:]:
        f, v = fv.split(':')
        idx.append(int(f))
        values.append(int(v))

    return f0, f1, idx, values

results = []

with gzip.open('data/criteo_train.txt.gz', 'r') as f:
#with gzip.open('data/criteo_train_small.txt.gz', 'r') as f:
    for line in tqdm(f):    
        line = line.decode()
        split = line.split('|')
        id = int(split[0].strip())

        label = None
        propensity = None
        features = None

        if len(split) == 4:
            l = split[1]
            assert l.startswith('l')

            l = l.lstrip('l ').strip()
            if l == '0.999':
                label = 0
            elif l == '0.001':
                label = 1
            else:
                raise Exception('ololo')

            p = split[2]
            assert p.startswith('p')
            p = p.lstrip('p ').strip()
            propensity = float(p)

            features = split[3].lstrip('f ').strip()

            f0, f1, idx, val = parse_features(features)
        elif len(split) == 2:
            continue
#             features = split[1].lstrip('f ').strip()
#         else:
#             raise Exception('ololo')

        results.append((id, f0, f1, idx, val, propensity, label))

len(results)

df_train = pd.DataFrame(results, columns=['id', 'f0', 'f1', 'col_idx', 'values', 'propensity', 'click'])

import pickle 
with open('tmp/df_train_clicks.bin', 'wb') as f:
    pickle.dump(df_train, f)

df_train.click.mean()

!free -h

import pickle 
with open('tmp/df_train_clicks.bin', 'rb') as f:
    df_train = pickle.load(f)

!free -h

y = df_train.click.values.astype('uint8')

X_num = df_train[['f0', 'f1']].values.astype('float32')

with open('tmp/y.bin', 'wb') as f:
    pickle.dump(y, f)

with open('tmp/X_num.bin', 'wb') as f:
    pickle.dump(X_num, f)

columns = list(df_train.col_idx)
values = list(df_train['values'])

del df_train
import gc
gc.collect()

def to_csr(cols, vals, shape=74000):
    lens = [len(c) for c in cols]
    intptr = np.zeros((len(cols) + 1), dtype='uint32')
    intptr[1:] = lens
    intptr = intptr.cumsum()

    columns = np.concatenate(cols).astype('uint32')
    values = np.concatenate(vals).astype('uint8')

    return sp.csr_matrix((values, columns, intptr), shape=(len(cols), shape))

sp.save_npz('matrix.npz', X, compressed=False)

!ls -lh

from sklearn.model_selection import train_test_split

X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.3, random_state=1)

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score
from time import time

#for C in [0.01, 0.1, 0.5, 1, 5]:
for C in [0.1]:
    t0 = time()

    lr = LogisticRegression(penalty='l1', C=C, random_state=1)
    lr.fit(X_train, y_train)

    y_pred = lr.decision_function(X_val)
    auc = roc_auc_score(y_val, y_pred)

    print('C=%s, took %.3fs, auc=%.3f' % (C, time() - t0, auc))

with open('model_c01.bin', 'wb') as f:
    pickle.dump(lr, f)

import pickle

with open('model_c01.bin', 'rb') as f:
    lr = pickle.load(f)

(lr.coef_[0] == 0).mean()

X.shape

!zcat data/criteo_test_release.txt.gz | head -n 10

def read_test_data():
    f = gzip.open('data/criteo_test_release.txt.gz', 'r')

    for line in f:    
        line = line.decode()
        split = line.split(' |f ')
        if len(split) != 2:
            continue

        id = int(split[0].strip())
        features = split[1].lstrip('f ').strip()
        yield (id, parse_features(features))

!zcat data/criteo_test_release.txt.gz | wc -l

from itertools import groupby




def sigmoid_percent(x):
    s = 1 / (1 + np.exp(-x))
    return (s * 100).round(2)

def to_prediction_str(id, preds):
    res = ['%d:%0.2f' % (i, p) for (i, p) in enumerate(preds)]
    return '%d;%s' % (id, ','.join(res))

test_data = read_test_data()
test_groups = groupby(test_data, key=lambda x: x[0])

7087738it groups

f_out = open('pred.txt', 'w')

for gid, group in tqdm(test_groups):
    cols = []
    vals = []

    for _, (f1, f2, col, val) in group:
        cols.append(col)
        vals.append(val)

    X_test = to_csr(cols, vals)
    pred = lr.decision_function(X_test)
    pred = sigmoid_percent(pred)
    
    pred_str = to_prediction_str(gid, pred)
    
    f_out.write(pred_str)
    f_out.write('\n')

f_out.flush()
f_out.close()

import crowdai
challenge = crowdai.Challenge("CriteoAdPlacementNIPS2017", 'd671d30799fa215f63a5cb5049983c79')

scores = challenge.submit('pred.txt.gz')

scores

scores
