==> criteo-nips-2017/02-cv.ipynb <==
import gzip
import pickle

import pandas as pd
import numpy as np
import scipy.sparse as sp

from tqdm import tqdm_notebook as tqdm

train_input = gzip.open('data/criteo_train.txt.gz', 'r')

files = [
    open('data/train_0.txt', 'w'),
    open('data/train_1.txt', 'w'),
    open('data/train_2.txt', 'w'),
    open('data/train_3.txt', 'w'),
]

for line in tqdm(train_input):
    line = line.decode()
    split = line.split('|')
    id = int(split[0].strip())
    fold = hash(id) % 4
    files[fold].write(line)

for f in files:
    f.flush()
    f.close()

173979748it

from collections import namedtuple
Line = namedtuple('Line', ['id', 'f0', 'f1', 'idx', 'val'])
LabeledLine = namedtuple('LabeledLine', ['id', 'f0', 'f1', 'idx', 'val', 'propensity', 'label'])

def parse_features(s):
    split = s.split(' ')
    f0 = split[0]
    assert f0.startswith('0:')
    f0 = int(f0[2:])

    f1 = split[1]
    assert f1.startswith('1:')
    f1 = int(f1[2:])

    idx = []
    values = []
    
    for fv in split[2:]:
        f, v = fv.split(':')
        idx.append(int(f) - 2)
        values.append(int(v))

    return f0, f1, idx, values

def read_train(fname):
    if fname.endswith('.gz'):
        f = gzip.open(fname, 'r')
        f = map(bytes.decode, f)
    else:
        f = open(fname, 'r')

    for line in f:
        split = line.split('|')
        id = int(split[0].strip())

        label = None
        propensity = None
        features = None

        if len(split) == 4:
            l = split[1]
            assert l.startswith('l')

            l = l.lstrip('l ').strip()
            if l == '0.999':
                label = 0
            elif l == '0.001':
                label = 1
            else:
                raise Exception('ololo')

            p = split[2]
            assert p.startswith('p')
            p = p.lstrip('p ').strip()
            propensity = float(p)

            features = split[3].lstrip('f ').strip()

            f0, f1, idx, val = parse_features(features)
            idx = np.array(idx, dtype=np.uint32)
            val = np.array(val, dtype=np.uint8)
            yield LabeledLine(id, f0, f1, idx, val, propensity, label)
        elif len(split) == 2:
            pass

import itertools
it0 = read_train('data/train_0.txt')
it1 = read_train('data/train_1.txt')
it2 = read_train('data/train_2.txt')
it_train = itertools.chain(it0, it1, it2)


df_train = []

for line in tqdm(it_train):
    df_train.append(line)

df_train = pd.DataFrame(df_train)

with open('tmp/df_train.bin', 'wb') as f:
    pickle.dump(df_train, f)

it_val = read_train('data/train_3.txt')

df_val = []

for line in tqdm(it_val):
    df_val.append(line)

df_val = pd.DataFrame(df_val)

with open('tmp/df_val.bin', 'wb') as f:
    pickle.dump(df_val, f)

Read data

import pickle

import pandas as pd
import numpy as np
import scipy.sparse as sp

with open('tmp/df_train.bin', 'rb') as f:
    df_train = pickle.load(f)

with open('tmp/df_val.bin', 'rb') as f:
    df_val = pickle.load(f)

def to_csr(cols, vals, shape=74000):
    lens = [len(c) for c in cols]
    intptr = np.zeros((len(cols) + 1), dtype='uint32')
    intptr[1:] = lens
    intptr = intptr.cumsum()

    columns = np.concatenate(cols).astype('uint32')
    values = np.concatenate(vals).astype('uint8')

    return sp.csr_matrix((values, columns, intptr), shape=(len(cols), shape))

X_train = to_csr(list(df_train.idx), list(df_train.val))
X_val = to_csr(list(df_val.idx), list(df_val.val))

sp.save_npz('tmp/X_train_sparse.npz', X_train, compressed=False)
sp.save_npz('tmp/X_val_sparse.npz', X_val, compressed=False)

y_train = df_train.label.values.astype('uint8')
y_val = df_val.label.values.astype('uint8')

prospensity_val = df_val.propensity.astype('float32')

X_num_train = df_train[['f0', 'f1']].values.astype('uint16')
X_num_val = df_val[['f0', 'f1']].values.astype('uint16')

np.save('tmp/y_train.npy', y_train)
np.save('tmp/y_val.npy', y_val)

np.save('tmp/X_num_train.npy', X_num_train)
np.save('tmp/X_num_val.npy', X_num_val)

np.save('tmp/prospensity_val.npy', prospensity_val)

Restart

import pickle

import pandas as pd
import numpy as np
import scipy.sparse as sp

X_train = sp.load_npz('tmp/X_train_sparse.npz')
X_val = sp.load_npz('tmp/X_val_sparse.npz')

y_train = np.load('tmp/y_train.npy', )
y_val = np.load('tmp/y_val.npy', )

X_num_train = np.load('tmp/X_num_train.npy', )
X_num_val = np.load('tmp/X_num_val.npy', )

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import roc_auc_score
from time import time

cnt = (X_train > 0).sum(axis=0)
cnt = np.asarray(cnt)[0]
mask = cnt >= 10

(cnt >= 10).mean()

X_train_f = X_train[:, mask]
X_val_f = X_val[:, mask]

X_train_f, X_val_f

log = open('exp.log', 'w')

for C in [0.05, 0.1, 0.5, 1]:
    t0 = time()

    lr = LogisticRegression(penalty='l1', C=C, random_state=1)
    lr.fit(X_train_f, y_train)

    y_pred = lr.decision_function(X_val_f)
    auc = roc_auc_score(y_val, y_pred)

    print('C=%s, took %.3fs, auc=%.3f' % (C, time() - t0, auc))
    log.write('C=%s, took %.3fs, auc=%.3f' % (C, time() - t0, auc))
    log.write('\n')
    log.flush()

;

log.flush()
log.close()
