==> criteo-nips-2017/04-metric.ipynb <==
import pandas as pd
import numpy as np
import scipy.sparse as sp

from sklearn.metrics import roc_auc_score
from tqdm import tqdm_notebook as tqdm

import competition_utils as u

X_train = sp.load_npz('tmp/X_train_sparse.npz')
X_val = sp.load_npz('tmp/X_val_sparse.npz')

y_train = np.load('tmp/y_train.npy', )
y_val = np.load('tmp/y_val.npy', )

y_train = y_train.astype(np.float32)
y_val = y_val.astype(np.float32)

prospensity_val = np.load('tmp/prospensity_val.npy')

import ftrl

model = ftrl.FtrlProximal(alpha=0.1, beta=1, l1=75, l2=25)

for i in tqdm(range(20)):
    model.fit(X_train, y_train)
    y_pred = model.predict(X_val)
    auc = roc_auc_score(y_val, y_pred)
    print(auc)

it_val = u.read_grouped('data/train_3.txt')

df_full_val = []

limit = 500000

cnt = 0

for id, group in tqdm(it_val, total=limit):
    df_full_val.extend(list(group))
    cnt = cnt + 1

    if cnt > limit:
        break

df_full_val = pd.DataFrame(df_full_val)

import pickle

with open('./tmp/df_full_val_500k.bin', 'rb') as f:
    df_full_val = pickle.load(f)

cols = list(df_full_val.idx)
vals = list(df_full_val.val)

X_val_full = u.to_csr(cols, vals)

%%time

df_full_val['predict'] = model.predict(X_val_full)

from numba import njit

prop_series = df_full_val.propensity.dropna()
num_groups = len(prop_series)

group_indptr = np.zeros(num_groups + 1, dtype='uint32')
group_indptr[:num_groups] = prop_series.index.values
group_indptr[-1] = len(df_full_val)

groups = df_full_val.id.values[group_indptr[:-1]]

prospensities = prop_series.values.astype('float32')
clicks = df_full_val.label.dropna().values

model_predictions = df_full_val.predict.values

@njit
def compute_IPS(model_predictions, group_indptr, prospensities, clicks):
    nom_sum = 0.0

    num_pos = 0
    num_neg = 0

    n = len(clicks)
    for i in range(n):
        label = clicks[i]

        if label == 0:
            num_neg = num_neg + 1
            continue

        num_pos = num_pos + 1
        prosp = prospensities[i]

        group_start = group_indptr[i]
        group_end = group_indptr[i + 1]
        scores = model_predictions[group_start:group_end]

        best = scores.max()
        offset_scores = scores - best

        probs = np.exp(offset_scores)
        prob = probs[0] / probs.sum()

        weight = prosp * prob 
        nom_sum = nom_sum + weight

    denom = num_pos + 10 * num_neg
    return 10000 * nom_sum / denom        

@njit
def argmax_increase(model_predictions, group_indptr, inc):
    out = model_predictions.copy()

    n_groups = len(group_indptr) - 1

    for i in range(n_groups):
        group_start = group_indptr[i]
        group_end = group_indptr[i + 1]
        
        scores = model_predictions[group_start:group_end]
    
        m = group_start + np.argmax(scores)
        out[m] = out[m] + inc

    return out

%%time
compute_IPS(model_predictions, group_indptr, prospensities, clicks)

compute_IPS(pred_inc, group_indptr, prospensities, clicks)

def shifted_scaled_sigmoid(x, shift=1, scale=1000000):
    s = 1 / (1 + np.exp(-x + shift))
    return (s * scale).round(2)

shift = 1.1875
scale = 850100
pred_scaled = shifted_scaled_sigmoid(model_predictions, shift=shift, scale=scale)

pred_inc = argmax_increase(pred_scaled, group_indptr, 1)

compute_IPS(pred_inc, group_indptr, prospensities, clicks)

from scipy.optimize.optimize import fmin

def f_optimize(x):
    pred = shifted_scaled_sigmoid(model_predictions, shift=x[0], scale=x[1])
    ips = compute_IPS(pred, group_indptr, prospensities, clicks)
    return -ips

x0 = np.array([1, 1000000])
#x0 = np.array([0, 100])

fmin(f_optimize, x0)

shift, scale = _
shift, scale

shift = 1.1875
scale = 850100


pred = shifted_scaled_sigmoid(model_predictions, shift=shift, scale=scale)
ips = compute_IPS(pred, group_indptr, prospensities, clicks)
ips

it_test = read_train('data/criteo_test_release.txt.gz')
test_groups = groupby(it_test, key=lambda x: x.id)

f_out = open('pred_ftrl.txt', 'w')

for gid, group in tqdm(test_groups, total=7087738):
    cols = []
    vals = []

    for line in group:
        cols.append(line.idx)
        vals.append(line.val)

    X_val = to_csr(cols, vals)
    pred = model_full.predict(X_val)
    pred = sigmoid_percent(pred)

    pred_str = to_prediction_str(gid, pred)
    
    f_out.write(pred_str)
    f_out.write('\n')

f_out.flush()
f_out.close()

!gzip pred_ftrl.txt

import crowdai
challenge = crowdai.Challenge("CriteoAdPlacementNIPS2017", 'd671d30799fa215f63a5cb5049983c79')

scores = challenge.submit('pred_ftrl.txt.gz')

scores

scores = challenge.submit('pred_ftrl.txt.gz')

scores
