==> criteo-nips-2017/05-ftrl-hash.ipynb <==
import pickle
import gzip

import pandas as pd
import numpy as np
import scipy.sparse as sp

from tqdm import tqdm_notebook as tqdm
from sklearn.metrics import roc_auc_score

import competition_utils as u

import numba as nb

X_train = sp.load_npz('tmp/X_train_sparse.npz')
X_val = sp.load_npz('tmp/X_val_sparse.npz')

y_train = np.load('tmp/y_train.npy', )
y_val = np.load('tmp/y_val.npy', )

y_train = y_train.astype(np.float32)
y_val = y_val.astype(np.float32)


X_num_train = np.load('tmp/X_num_train.npy', )
X_num_val = np.load('tmp/X_num_val.npy', )

@nb.vectorize([nb.int32(nb.int32)], nopython=True, target='parallel')
def hashed(x):
    x = ((x >> 16) ^ x) * 0x45d9f3b
    x = ((x >> 16) ^ x) * 0x45d9f3b
    x = (x >> 16) ^ x
    return x

@nb.vectorize([nb.int32(nb.int32, nb.int32)], nopython=True, target='parallel')
def nonneg_mod(i, mod):
    m = i % mod
    if m < 0:
        m = m + mod
    return m

def rehash_matrix(X):
    n, _ = X.shape
    D = 2 ** 22
    indices_hashed = nonneg_mod(hashed(X.indices) + 13 * hashed(X.data), D)
    data = np.ones_like(X.data, dtype=np.uint8)
    X_res = sp.csr_matrix((data, indices_hashed, X.indptr), shape=(n, D))
    return X_res

%%time
X_train = rehash_matrix(X_train)
X_val = rehash_matrix(X_val)

from sklearn.preprocessing import OneHotEncoder

X_num_train = X_num_train.astype('uint32')
X_num_train = np.hstack([X_num_train, (X_num_train[:, 0] * X_num_train[:, 1]).reshape(-1, 1)], )

X_num_val = X_num_val.astype('uint32')
X_num_val = np.hstack([X_num_val, (X_num_val[:, 0] * X_num_val[:, 1]).reshape(-1, 1)], )

ohe1 = OneHotEncoder(dtype='uint8')
X_num_ohe_train = ohe1.fit_transform(X_num_train)
X_num_ohe_val = ohe1.transform(X_num_val)

cnt_num = X_num_ohe_train.sum(axis=0)
cnt_num = np.asarray(cnt_num)[0]
mask_num = cnt_num >= 50

X_num_ohe_train = X_num_ohe_train[:, mask_num]
X_num_ohe_val = X_num_ohe_val[:, mask_num]

model_num = ftrl.FtrlProximal(alpha=0.5, beta=1, l1=1, l2=0)

for i in tqdm(range(30)):
    model_num.fit(X_num_ohe_train, y_train)
    y_pred = model_num.predict(X_num_ohe_val)
    auc = roc_auc_score(y_val, y_pred)
    print(i + 1, auc)

cnt = (X_train > 0).sum(axis=0)
cnt = np.asarray(cnt)[0]
mask = cnt >= 50

X_train = X_train[:, mask]
X_val = X_val[:, mask]

X_train = sp.hstack([X_num_ohe_train, X_train, ], format='csr')
X_val = sp.hstack([X_num_ohe_val, X_val, ], format='csr')

import ftrl

for i in tqdm(range(30)):
    model.fit(X_train, y_train)
    y_pred = model.predict(X_val)
    auc = roc_auc_score(y_val, y_pred)
    print(i + 1, auc)

for i in tqdm(range(30)):
    model.fit(X_train, y_train)
    y_pred = model.predict(X_val)
    auc = roc_auc_score(y_val, y_pred)
    print(i + 1, auc)

X = sp.vstack([X_train, X_val])
y = np.concatenate([y_train, y_val])

Full model

%%time

model_full = ftrl.FtrlProximal(alpha=0.1, beta=1, l1=75, l2=25)
model_full.fit(X, y, num_passes=22)

shift = 1.1875
scale = 850100

def shifted_scaled_sigmoid(x, shift=0, scale=1):
    s = 1 / (1 + np.exp(-x + shift))
    return (s * scale).round(2)

it_test = u.read_grouped('data/criteo_test_release.txt.gz')

f_out = open('pred_ftrl2.txt', 'w')

for gid, group in tqdm(it_test, total=7087738):
    cols = []
    vals = []

    for line in group:
        cols.append(line.idx)
        vals.append(line.val)

    X_val = u.to_csr(cols, vals)

    pred = model_full.predict(X_val)
    pred = shifted_scaled_sigmoid(pred, shift, scale)

    pred_str = u.to_prediction_str(gid, pred)
    
    f_out.write(pred_str)
    f_out.write('\n')

f_out.flush()
f_out.close()

!gzip pred_ftrl2.txt

import crowdai
challenge = crowdai.Challenge("CriteoAdPlacementNIPS2017", 'd671d30799fa215f63a5cb5049983c79')

scores = challenge.submit('pred_ftrl2.txt.gz')

scores

scores
